+++ src/commands/refactor.rs
@@ -1,5 +1,5 @@
 use clap::{Args, Subcommand};
-use homeboy::core::code_audit::{AuditFinding, CodeAuditResult};
+use homeboy::core::code_audit::AuditFinding;
 use homeboy::core::engine::execution_context::{self, ResolveOptions};
 use homeboy::core::refactor::{
     self, auto, AddResult, MoveResult, RenameContext, RenameScope, RenameSpec, RenameTargeting,
@@ -11,6 +11,8 @@ use super::utils::args::{BaselineArgs, PositionalComponentArgs, SettingArgs, Wri
 use crate::commands::CmdResult;
 
 mod autofix_commit;
+mod operations_command;
+mod transform_command;
 
 #[derive(Args)]
 #[command(args_conflicts_with_subcommands = true, subcommand_negates_reqs = true)]
@@ -313,7 +315,7 @@ pub fn run(args: RefactorArgs, _global: &crate::commands::GlobalArgs) -> CmdResu
             to,
             target,
             write_mode,
-        }) => run_add(
+        }) => operations_command::run_add(
             from_audit.as_deref(),
             import.as_deref(),
             to.as_deref(),
@@ -330,7 +332,7 @@ pub fn run(args: RefactorArgs, _global: &crate::commands::GlobalArgs) -> CmdResu
             write_mode,
         }) => {
             if let Some(file_path) = file {
-                run_move_file(&file_path, &to, &target, write_mode.write)
+                operations_command::run_move_file(&file_path, &to, &target, write_mode.write)
             } else if let Some(from_path) = from {
                 if item.is_empty() {
                     return Err(homeboy::core::Error::validation_invalid_argument(
@@ -344,7 +346,7 @@ pub fn run(args: RefactorArgs, _global: &crate::commands::GlobalArgs) -> CmdResu
                         ]),
                     ));
                 }
-                run_move(&item, &from_path, &to, &target, write_mode.write)
+                operations_command::run_move(&item, &from_path, &to, &target, write_mode.write)
             } else {
                 Err(homeboy::core::Error::validation_invalid_argument(
                     "from",
@@ -364,7 +366,7 @@ pub fn run(args: RefactorArgs, _global: &crate::commands::GlobalArgs) -> CmdResu
             definition,
             target,
             write_mode,
-        }) => run_propagate(
+        }) => operations_command::run_propagate(
             &struct_name,
             definition.as_deref(),
             &target,
@@ -379,7 +381,7 @@ pub fn run(args: RefactorArgs, _global: &crate::commands::GlobalArgs) -> CmdResu
             full_match_details,
             target,
             write_mode,
-        }) => run_transform(
+        }) => transform_command::run_transform(
             &find,
             &replace,
             &files,
@@ -394,7 +396,7 @@ pub fn run(args: RefactorArgs, _global: &crate::commands::GlobalArgs) -> CmdResu
             strategy,
             target,
             write_mode,
-        }) => run_decompose(&file, &strategy, &target, write_mode.write),
+        }) => operations_command::run_decompose(&file, &strategy, &target, write_mode.write),
     }
 }
 
@@ -993,635 +995,6 @@ fn parse_rename_variants(values: &[String]) -> homeboy::core::Result<Vec<(String
             Ok((from.to_string(), to.to_string()))
         })
         .collect()
-}
-
-fn run_add(
-    from_audit: Option<&str>,
-    import: Option<&str>,
-    to: Option<&str>,
-    target: &RefactorTargetArgs,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    // Mode 1: From audit JSON
-    if let Some(audit_source) = from_audit {
-        return run_add_from_audit(audit_source, write);
-    }
-
-    // Mode 2: Explicit import addition
-    if let Some(import_line) = import {
-        let destination = to.ok_or_else(|| {
-            homeboy::core::Error::validation_invalid_argument(
-                "to",
-                "--to is required when using --import",
-                None,
-                Some(vec![
-                    "homeboy refactor add --import \"use serde::Serialize;\" --to \"src/**/*.rs\""
-                        .to_string(),
-                ]),
-            )
-        })?;
-
-        let targets = target.resolve_targets()?;
-        return run_across_targets("add", targets, |component_id, path| {
-            run_add_import(import_line, destination, component_id, path, write)
-        });
-    }
-
-    // Neither mode specified
-    Err(homeboy::core::Error::validation_invalid_argument(
-        "add",
-        "Specify either --from-audit or --import with --to",
-        None,
-        Some(vec![
-            "homeboy refactor add --from-audit @audit.json".to_string(),
-            "homeboy refactor add --import \"use serde::Serialize;\" --to \"src/**/*.rs\""
-                .to_string(),
-        ]),
-    ))
-}
-
-fn run_add_from_audit(source: &str, write: bool) -> CmdResult<RefactorOutput> {
-    // Parse audit JSON from @file, stdin (-), file path, or inline string.
-    // Auto-detect bare file paths (same pattern as docs generate --from-audit).
-    let effective_source = if !source.starts_with('{')
-        && !source.starts_with('[')
-        && source != "-"
-        && !source.starts_with('@')
-        && std::path::Path::new(source).exists()
-    {
-        format!("@{}", source)
-    } else {
-        source.to_string()
-    };
-
-    let json_content = crate::commands::merge_json_sources(Some(&effective_source), &[])?;
-
-    // Parse audit result — handle both envelope and raw formats.
-    // The envelope format wraps the audit in a "data" field (from `homeboy --format json audit`).
-    let audit: CodeAuditResult = if let Some(data) = json_content.get("data") {
-        serde_json::from_value(data.clone())
-    } else {
-        serde_json::from_value(json_content)
-    }
-    .map_err(|e| {
-        homeboy::core::Error::validation_invalid_json(
-            e,
-            Some("parse audit result for refactor add".to_string()),
-            Some(
-                "Input must be output from `homeboy audit <component>`. \
-                 Save it with: homeboy --format json audit <component> > audit.json"
-                    .to_string(),
-            ),
-        )
-    })?;
-
-    let fix_result = refactor::fixes_from_audit(&audit, write)?;
-
-    let exit_code = if fix_result.total_insertions > 0 {
-        1
-    } else {
-        0
-    };
-
-    homeboy::log_status!(
-        "refactor",
-        "{} fix(es) across {} file(s){}",
-        fix_result.total_insertions,
-        fix_result.fixes.len(),
-        if write {
-            format!(" — {} written", fix_result.files_modified)
-        } else {
-            " (dry run)".to_string()
-        }
-    );
-
-    Ok((
-        RefactorOutput::AddFromAudit {
-            source_path: audit.source_path,
-            fix_result,
-            dry_run: !write,
-        },
-        exit_code,
-    ))
-}
-
-fn run_add_import(
-    import_line: &str,
-    target: &str,
-    component_id: Option<&str>,
-    path: Option<&str>,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let result = refactor::add_import(import_line, target, component_id, path, write)?;
-
-    let exit_code = if result.total_insertions > 0 { 1 } else { 0 };
-
-    homeboy::log_status!(
-        "refactor",
-        "{} file(s) to update with '{}'{}",
-        result.total_insertions,
-        import_line,
-        if write {
-            format!(" — {} written", result.files_modified)
-        } else {
-            " (dry run)".to_string()
-        }
-    );
-
-    Ok((
-        RefactorOutput::AddImport {
-            import: import_line.to_string(),
-            target: target.to_string(),
-            result,
-            dry_run: !write,
-        },
-        exit_code,
-    ))
-}
-
-fn run_move(
-    items: &[String],
-    from: &str,
-    to: &str,
-    target: &RefactorTargetArgs,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let targets = target.resolve_targets()?;
-    run_across_targets("move", targets, |component_id, path| {
-        run_move_single(items, from, to, component_id, path, write)
-    })
-}
-
-fn run_move_single(
-    items: &[String],
-    from: &str,
-    to: &str,
-    component_id: Option<&str>,
-    path: Option<&str>,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let root = refactor::move_items::resolve_root(component_id, path)?;
-
-    if write {
-        homeboy::core::engine::undo::UndoSnapshot::capture_and_save(
-            &root,
-            "refactor move",
-            [from, to],
-        );
-    }
-
-    let item_refs: Vec<&str> = items.iter().map(|s| s.as_str()).collect();
-    let result = refactor::move_items(&item_refs, from, to, &root, write)?;
-
-    let exit_code = if result.items_moved.is_empty() { 1 } else { 0 };
-
-    homeboy::log_status!(
-        "refactor",
-        "{} item(s) from {} → {}{}",
-        result.items_moved.len(),
-        from,
-        to,
-        if write {
-            " (applied)".to_string()
-        } else {
-            " (dry run)".to_string()
-        }
-    );
-
-    for item in &result.items_moved {
-        homeboy::log_status!(
-            "move",
-            "{} {:?} (lines {}-{})",
-            item.name,
-            item.kind,
-            item.source_lines.0,
-            item.source_lines.1
-        );
-    }
-
-    for test in &result.tests_moved {
-        homeboy::log_status!(
-            "move",
-            "test {} (lines {}-{})",
-            test.name,
-            test.source_lines.0,
-            test.source_lines.1
-        );
-    }
-
-    if result.imports_updated > 0 {
-        homeboy::log_status!(
-            "move",
-            "{} import reference(s) updated across codebase",
-            result.imports_updated
-        );
-    }
-
-    for warning in &result.warnings {
-        homeboy::log_status!("warning", "{}", warning);
-    }
-
-    Ok((RefactorOutput::Move { result }, exit_code))
-}
-
-fn run_move_file(
-    file: &str,
-    to: &str,
-    target: &RefactorTargetArgs,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let targets = target.resolve_targets()?;
-    run_across_targets("move_file", targets, |component_id, path| {
-        run_move_file_single(file, to, component_id, path, write)
-    })
-}
-
-fn run_move_file_single(
-    file: &str,
-    to: &str,
-    component_id: Option<&str>,
-    path: Option<&str>,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let root = refactor::move_items::resolve_root(component_id, path)?;
-
-    if write {
-        homeboy::core::engine::undo::UndoSnapshot::capture_and_save(
-            &root,
-            "refactor move --file",
-            [file, to],
-        );
-    }
-
-    let result = refactor::move_items::move_file(file, to, &root, write)?;
-
-    let exit_code = if result.imports_updated > 0 || result.mod_declarations_updated {
-        0
-    } else {
-        1
-    };
-
-    homeboy::log_status!(
-        "refactor",
-        "move {} → {}{}",
-        file,
-        to,
-        if write { " (applied)" } else { " (dry run)" }
-    );
-    homeboy::log_status!(
-        "move",
-        "{} import(s) rewritten across {} file(s)",
-        result.imports_updated,
-        result.caller_files_modified.len()
-    );
-    if result.mod_declarations_updated {
-        homeboy::log_status!("move", "mod.rs declarations updated");
-    }
-    for warning in &result.warnings {
-        homeboy::log_status!("warning", "{}", warning);
-    }
-
-    Ok((RefactorOutput::MoveFile { result }, exit_code))
-}
-
-// ============================================================================
-// Propagate (add missing fields to struct instantiations)
-// ============================================================================
-
-fn run_propagate(
-    struct_name: &str,
-    definition_file: Option<&str>,
-    target: &RefactorTargetArgs,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let targets = target.resolve_targets()?;
-    run_across_targets("propagate", targets, |component_id, path| {
-        run_propagate_single(struct_name, definition_file, component_id, path, write)
-    })
-}
-
-fn run_propagate_single(
-    struct_name: &str,
-    definition_file: Option<&str>,
-    component_id: Option<&str>,
-    path: Option<&str>,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let root = refactor::move_items::resolve_root(component_id, path)?;
-
-    // Capture undo snapshot before writes
-    let config = refactor::PropagateConfig {
-        struct_name,
-        definition_file,
-        root: &root,
-        write: false, // dry-run first if we need undo
-    };
-
-    if write {
-        // Dry-run to discover affected files for the undo snapshot
-        let preview = refactor::propagate(&config)?;
-        let affected_files: Vec<&str> = preview.edits.iter().map(|e| e.file.as_str()).collect();
-        homeboy::core::engine::undo::UndoSnapshot::capture_and_save(
-            &root,
-            "refactor propagate",
-            affected_files,
-        );
-    }
-
-    // Run the actual propagation (with write mode as requested)
-    let write_config = refactor::PropagateConfig {
-        struct_name,
-        definition_file,
-        root: &root,
-        write,
-    };
-    let result = refactor::propagate(&write_config)?;
-
-    // Log results to stderr
-    homeboy::log_status!(
-        "propagate",
-        "{} instantiation(s) found, {} need fixes, {} edit(s){}",
-        result.instantiations_found,
-        result.instantiations_needing_fix,
-        result.edits.len(),
-        if write {
-            if result.applied {
-                " (applied)".to_string()
-            } else {
-                " (nothing to apply)".to_string()
-            }
-        } else {
-            " (dry run)".to_string()
-        }
-    );
-
-    for edit in &result.edits {
-        homeboy::log_status!("edit", "{}:{} — {}", edit.file, edit.line, edit.description);
-    }
-
-    let exit_code = if result.edits.is_empty() { 0 } else { 1 };
-
-    Ok((
-        RefactorOutput::Propagate {
-            result,
-            dry_run: !write,
-        },
-        exit_code,
-    ))
-}
-
-// ============================================================================
-// Transform
-// ============================================================================
-
-fn run_transform(
-    find: &str,
-    replace: &str,
-    files: &str,
-    context: &str,
-    full_match_details: bool,
-    target: &RefactorTargetArgs,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let targets = target.resolve_targets()?;
-    run_across_targets("transform", targets, |component_id, path| {
-        run_transform_single(
-            find,
-            replace,
-            files,
-            context,
-            full_match_details,
-            component_id,
-            path,
-            write,
-        )
-    })
-}
-
-fn run_transform_single(
-    find: &str,
-    replace: &str,
-    files: &str,
-    context: &str,
-    full_match_details: bool,
-    component_id: Option<&str>,
-    path: Option<&str>,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let root = refactor::move_items::resolve_root(component_id, path)?;
-    let set_name = "ad-hoc";
-    let set = refactor::ad_hoc_transform(find, replace, files, context);
-
-    // Report what we're about to do
-    homeboy::log_status!(
-        "transform",
-        "{} ({} rule{})",
-        set_name,
-        set.rules.len(),
-        if set.rules.len() == 1 { "" } else { "s" }
-    );
-
-    if !set.description.is_empty() {
-        homeboy::log_status!("info", "{}", set.description);
-    }
-
-    if write {
-        // Dry-run to discover affected files for the undo snapshot
-        if let Ok(preview) = refactor::apply_transforms_with_options(
-            &root,
-            set_name,
-            &set,
-            false,
-            None,
-            refactor::TransformOptions::default(),
-        ) {
-            let affected_files: std::collections::HashSet<String> =
-                preview.modified_files.iter().cloned().collect();
-            homeboy::core::engine::undo::UndoSnapshot::capture_and_save(
-                &root,
-                "refactor transform",
-                &affected_files,
-            );
-        }
-    }
-
-    // Apply transforms
-    let result = refactor::apply_transforms_with_options(
-        &root,
-        set_name,
-        &set,
-        write,
-        None,
-        refactor::TransformOptions {
-            match_detail_limit: if full_match_details {
-                None
-            } else {
-                Some(refactor::DEFAULT_MATCH_DETAIL_LIMIT)
-            },
-        },
-    )?;
-
-    // Report results to stderr
-    for rule_result in &result.rules {
-        if rule_result.matches.is_empty() {
-            homeboy::log_status!("skip", "{}: no matches", rule_result.id);
-            continue;
-        }
-
-        homeboy::log_status!(
-            "rule",
-            "{}: {} replacement{}",
-            rule_result.id,
-            rule_result.replacement_count,
-            if rule_result.replacement_count == 1 {
-                ""
-            } else {
-                "s"
-            }
-        );
-
-        for m in &rule_result.matches {
-            homeboy::log_status!("  match", "{}:{}", m.file, m.line);
-            if !m.before.is_empty() {
-                homeboy::log_status!("  -", "{}", m.before.trim());
-                homeboy::log_status!("  +", "{}", m.after.trim());
-            }
-        }
-
-        if rule_result.matches_truncated {
-            homeboy::log_status!(
-                "  omitted",
-                "{} additional match detail{} omitted (use --full-match-details to include all)",
-                rule_result.omitted_match_count,
-                if rule_result.omitted_match_count == 1 {
-                    ""
-                } else {
-                    "s"
-                }
-            );
-        }
-    }
-
-    // Summary
-    if result.total_replacements == 0 {
-        homeboy::log_status!("result", "No matches found");
-    } else if write {
-        homeboy::log_status!(
-            "result",
-            "{} replacement{} applied across {} file{}",
-            result.total_replacements,
-            if result.total_replacements == 1 {
-                ""
-            } else {
-                "s"
-            },
-            result.total_files,
-            if result.total_files == 1 { "" } else { "s" },
-        );
-    } else {
-        homeboy::log_status!(
-            "result",
-            "{} replacement{} across {} file{} (dry-run, use --write to apply)",
-            result.total_replacements,
-            if result.total_replacements == 1 {
-                ""
-            } else {
-                "s"
-            },
-            result.total_files,
-            if result.total_files == 1 { "" } else { "s" },
-        );
-    }
-
-    let exit_code = if result.total_replacements == 0 { 1 } else { 0 };
-    Ok((RefactorOutput::Transform { result }, exit_code))
-}
-
-fn run_decompose(
-    file: &str,
-    strategy: &str,
-    target: &RefactorTargetArgs,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let targets = target.resolve_targets()?;
-    run_across_targets("decompose", targets, |component_id, path| {
-        run_decompose_single(file, strategy, component_id, path, write)
-    })
-}
-
-fn run_decompose_single(
-    file: &str,
-    strategy: &str,
-    component_id: Option<&str>,
-    path: Option<&str>,
-    write: bool,
-) -> CmdResult<RefactorOutput> {
-    let root = refactor::move_items::resolve_root(component_id, path)?;
-    let plan = refactor::build_plan(file, &root, strategy)?;
-
-    if write {
-        let affected: Vec<&str> = std::iter::once(file)
-            .chain(plan.groups.iter().map(|g| g.suggested_target.as_str()))
-            .collect();
-        homeboy::core::engine::undo::UndoSnapshot::capture_and_save(
-            &root,
-            "refactor decompose",
-            &affected,
-        );
-    }
-
-    let move_results = refactor::apply_plan(&plan, &root, write)?;
-    let groups_applied = move_results
-        .iter()
-        .filter(|result| !result.items_moved.is_empty())
-        .count();
-
-    homeboy::log_status!(
-        "decompose",
-        "{} group(s) planned for {}{}",
-        plan.groups.len(),
-        file,
-        if write { " (applied)" } else { " (dry run)" }
-    );
-
-    for group in &plan.groups {
-        homeboy::log_status!(
-            "decompose",
-            "{} -> {} ({} item(s))",
-            group.name,
-            group.suggested_target,
-            group.item_names.len()
-        );
-    }
-
-    if !plan.warnings.is_empty() {
-        for warning in &plan.warnings {
-            homeboy::log_status!("warning", "{}", warning);
-        }
-    }
-
-    if !plan.projected_audit_impact.likely_findings.is_empty() {
-        for finding in &plan.projected_audit_impact.likely_findings {
-            homeboy::log_status!("impact", "{}", finding);
-        }
-    }
-
-    homeboy::log_status!(
-        "decompose",
-        "{} move group(s) {}",
-        groups_applied,
-        if write { "applied" } else { "planned" }
-    );
-
-    Ok((
-        RefactorOutput::Decompose {
-            plan,
-            move_results,
-            dry_run: !write,
-            applied: write,
-        },
-        0,
-    ))
 }
 
 #[cfg(test)]
